+++ clients/staking/tests/unit/restapi/api_on_chain_yields_test.go
@@ -347,7 +347,7 @@ func Test_binancestakingrestapi_OnChainYieldsAPIService(t *testing.T) {
 		mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			require.Equal(t, "/sapi/v1/onchain-yields/locked/subscriptionPreview", r.URL.Path)
 			require.Equal(t, "1", r.URL.Query().Get("projectId"))
-			require.Equal(t, fmt.Sprintf("%v", float32(1.0)), r.URL.Query().Get("amount"))
+			require.Equal(t, fmt.Sprintf("%v", float64(1.0)), r.URL.Query().Get("amount"))
 			w.Header().Set("Content-Type", "application/json")
 			_, _ = w.Write([]byte(mockedJSON))
 		}))
@@ -364,7 +364,7 @@ func Test_binancestakingrestapi_OnChainYieldsAPIService(t *testing.T) {
 			client.WithRestAPI(configuration),
 		)
 
-		resp, err := apiClient.RestApi.OnChainYieldsAPI.GetOnChainYieldsLockedSubscriptionPreview(context.Background()).ProjectId("1").Amount(float32(1.0)).Execute()
+		resp, err := apiClient.RestApi.OnChainYieldsAPI.GetOnChainYieldsLockedSubscriptionPreview(context.Background()).ProjectId("1").Amount(float64(1.0)).Execute()
 		require.NoError(t, err)
 		require.NotNil(t, resp)
 		require.Equal(
@@ -776,7 +776,7 @@ func Test_binancestakingrestapi_OnChainYieldsAPIService(t *testing.T) {
 		mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			require.Equal(t, "/sapi/v1/onchain-yields/locked/subscribe", r.URL.Path)
 			require.Equal(t, "1", r.URL.Query().Get("projectId"))
-			require.Equal(t, fmt.Sprintf("%v", float32(1.0)), r.URL.Query().Get("amount"))
+			require.Equal(t, fmt.Sprintf("%v", float64(1.0)), r.URL.Query().Get("amount"))
 			w.Header().Set("Content-Type", "application/json")
 			_, _ = w.Write([]byte(mockedJSON))
 		}))
@@ -793,7 +793,7 @@ func Test_binancestakingrestapi_OnChainYieldsAPIService(t *testing.T) {
 			client.WithRestAPI(configuration),
 		)
 
-		resp, err := apiClient.RestApi.OnChainYieldsAPI.SubscribeOnChainYieldsLockedProduct(context.Background()).ProjectId("1").Amount(float32(1.0)).Execute()
+		resp, err := apiClient.RestApi.OnChainYieldsAPI.SubscribeOnChainYieldsLockedProduct(context.Background()).ProjectId("1").Amount(float64(1.0)).Execute()
 		require.NoError(t, err)
 		require.NotNil(t, resp)
 		require.Equal(