+++ clients/staking/tests/unit/restapi/api_sol_staking_test.go
@@ -523,7 +523,7 @@ func Test_binancestakingrestapi_SolStakingAPIService(t *testing.T) {
 		}
 		mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			require.Equal(t, "/sapi/v1/sol-staking/sol/redeem", r.URL.Path)
-			require.Equal(t, fmt.Sprintf("%v", float32(1.0)), r.URL.Query().Get("amount"))
+			require.Equal(t, fmt.Sprintf("%v", float64(1.0)), r.URL.Query().Get("amount"))
 			w.Header().Set("Content-Type", "application/json")
 			_, _ = w.Write([]byte(mockedJSON))
 		}))
@@ -540,7 +540,7 @@ func Test_binancestakingrestapi_SolStakingAPIService(t *testing.T) {
 			client.WithRestAPI(configuration),
 		)
 
-		resp, err := apiClient.RestApi.SolStakingAPI.RedeemSol(context.Background()).Amount(float32(1.0)).Execute()
+		resp, err := apiClient.RestApi.SolStakingAPI.RedeemSol(context.Background()).Amount(float64(1.0)).Execute()
 		require.NoError(t, err)
 		require.NotNil(t, resp)
 		require.Equal(
@@ -659,7 +659,7 @@ func Test_binancestakingrestapi_SolStakingAPIService(t *testing.T) {
 		}
 		mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			require.Equal(t, "/sapi/v1/sol-staking/sol/stake", r.URL.Path)
-			require.Equal(t, fmt.Sprintf("%v", float32(1.0)), r.URL.Query().Get("amount"))
+			require.Equal(t, fmt.Sprintf("%v", float64(1.0)), r.URL.Query().Get("amount"))
 			w.Header().Set("Content-Type", "application/json")
 			_, _ = w.Write([]byte(mockedJSON))
 		}))
@@ -676,7 +676,7 @@ func Test_binancestakingrestapi_SolStakingAPIService(t *testing.T) {
 			client.WithRestAPI(configuration),
 		)
 
-		resp, err := apiClient.RestApi.SolStakingAPI.SubscribeSolStaking(context.Background()).Amount(float32(1.0)).Execute()
+		resp, err := apiClient.RestApi.SolStakingAPI.SubscribeSolStaking(context.Background()).Amount(float64(1.0)).Execute()
 		require.NoError(t, err)
 		require.NotNil(t, resp)
 		require.Equal(